requirements.txt
pytest
pytz
fastapi
httpx
uvicorn
segment-analytics-python
python-dotenv
starlette
envclasses
./pypi/qdrant_analytics_events-1.0.13.tar.gz